Junos and VI model: support uint32 admin cost with preference add/subtract
Extend administrative cost (Juniper preference) from int to long to support the full uint32 range (0-4,294,967,295) required by Junos. Add support for `set policy-options policy-statement PS term TERM then preference add/subtract <p>`.

Changes:
- Update AbstractRoute._admin field from int to long with MAX_ADMIN_DISTANCE changed to 0xFFFFFFFFL
- Update all route subclasses, builders, and infrastructure to use long for admin cost
- Fix all route hashCode methods to use Long.hashCode(_admin) instead of casting to int
- Create AdministrativeCostExpr hierarchy (not LongExpr) with LiteralAdministrativeCost, IncrementAdministrativeCost(addend, max), and DecrementAdministrativeCost(subtrahend, min)
- Increment/Decrement classes require explicit min/max bounds (no defaults) to support per-vendor ranges
- Update SetAdministrativeCost to accept AdministrativeCostExpr instead of IntExpr
- Update Junos grammar to parse `then preference [add|subtract] <uint32>`
- Update PsThenPreference to support operators (ADD/SUBTRACT/SET) similar to PsThenLocalPreference
- Update ConfigurationBuilder extraction logic to handle preference operators
- Update TransferBDD BDD representation from 8 bits to 32 bits for admin cost
- Add TransferBDD support for Increment/Decrement using addClipping/subClipping
- Update question schema to use Schema.LONG for admin distance columns
- Fix BgpRoute to use Long type for DEFAULT_AD and parameter for type inference
- Add comprehensive unit tests for all new classes (equals, serialization, evaluation)
- Add extraction test for preference add/subtract parsing into PsThenPreference
- Add conversion test for PsThenPreference.applyTo() with all operators
- Add TransferBDD symbolic analysis tests for increment/decrement operations
- Update all test assertions and reference files for long type

Resolves https://github.com/batfish/batfish/issues/8808

---

Prompt:
```
Add support for Junos `set policy-options policy-statement PS term TERM then preference subtract <p>`. Refer to the Junos CLI reference in working/ -- there's probably an `add` as well?

We need to implement this fully: Juniper parsing and extraction, and vendor-independent model RoutingPolicy statement. We need to make sure that subtract and add are clipped to the valid range. And we need to support them in TransferBDD.
```

Further discussion:
- Expanded scope to change AbstractRoute._admin from int to long throughout entire VI model
- Created dedicated AdministrativeCostExpr hierarchy instead of using LongExpr for semantic correctness
- Increment/Decrement require explicit bounds to support per-vendor admin cost ranges (Juniper uses 0-4,294,967,295, most others use 0-255)
- Updated BDD representation to use 32 bits instead of 8 bits for full uint32 range support
- Used addClipping/subClipping in TransferBDD for proper BDD arithmetic
- Fixed all review comments including type inference issues, null checks, and test coverage

commit-id:d8a97815

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/OspfRoute.java
@@ -25,7 +25,7 @@ public abstract class OspfRoute extends AbstractRoute {
   protected OspfRoute(
       Prefix network,
       NextHop nextHop,
-      int admin,
+      long admin,
       long metric,
       long area,
       long tag,
